+++ tests/ColumnConsistencyTest.php
@@ -1,0 +1,59 @@
+<?php
+
+namespace LeagueTest\Csv;
+
+use League\Csv\ColumnConsistency;
+use League\Csv\Exception\InsertionException;
+use League\Csv\Exception\OutOfRangeException;
+use League\Csv\Writer;
+use PHPUnit\Framework\TestCase;
+use SplFileObject;
+use SplTempFileObject;
+
+/**
+ * @group validators
+ */
+class ColumnConsistencyTest extends TestCase
+{
+    private $csv;
+
+    public function setUp()
+    {
+        $this->csv = Writer::createFromFileObject(new SplTempFileObject());
+    }
+
+    public function tearDown()
+    {
+        $csv = new SplFileObject(__DIR__.'/data/foo.csv', 'w');
+        $csv->setCsvControl();
+        $csv->fputcsv(['john', 'doe', 'john.doe@example.com'], ',', '"');
+        $this->csv = null;
+    }
+
+    public function testAutoDetect()
+    {
+        try {
+            $expected = ['jane', 'jane.doe@example.com'];
+            $this->csv->addValidator((new ColumnConsistency())->autodetect(), 'consistency');
+            $this->csv->insertOne(['john', 'doe', 'john.doe@example.com']);
+            $this->csv->insertOne($expected);
+        } catch (InsertionException $e) {
+            $this->assertSame($e->getName(), 'consistency');
+            $this->assertEquals($e->getRecord(), ['jane', 'jane.doe@example.com']);
+        }
+    }
+
+    public function testColumnsCount()
+    {
+        $this->expectException(InsertionException::class);
+        $this->csv->addValidator((new ColumnConsistency())->columnsCount(3), 'consistency');
+        $this->csv->insertOne(['john', 'doe', 'john.doe@example.com']);
+        $this->csv->insertOne(['jane', 'jane.doe@example.com']);
+    }
+
+    public function testColumsCountTriggersException()
+    {
+        $this->expectException(OutOfRangeException::class);
+        (new ColumnConsistency())->columnsCount(-1);
+    }
+}